+++ nzshm_common/__init__.py
@@ -3,3 +3,7 @@ __email__ = 'nshm@gns.cri.nz'
 __version__ = '0.7.0'
 
 from .location import location
+
+# Common classes at the top level for convenience
+from .location.code_location import CodedLocation
+from .location.types import LatLon