+++ nzshm_common/grids/io.py
@@ -2,9 +2,12 @@ import base64
 import csv
 import io
 import zipfile
+from typing import List
+
+from nzshm_common.location.types import LatLon
 
 
-def latlon_from_base64_zip(b64_data: str) -> list:
+def latlon_from_base64_zip(b64_data: str) -> List[LatLon]:
     result = []
     byte_data = base64.b64decode(b64_data)
     data = io.BytesIO(byte_data)
@@ -12,5 +15,5 @@ def latlon_from_base64_zip(b64_data: str) -> list:
         plain_bytes = io.BytesIO(fz.read("data"))
         reader = csv.reader(io.StringIO(plain_bytes.getvalue().decode("utf-8", "ignore")))
         for row in reader:
-            result.append((float(row[1]), float(row[0])))
+            result.append(LatLon(float(row[1]), float(row[0])))
     return result