+++ nzshm_common/location/code_location.py
@@ -1,6 +1,9 @@
 import decimal
 from dataclasses import dataclass, field
 
+from nzshm_common.constants import DEFAULT_RESOLUTION
+from nzshm_common.location.types import LatLon
+
 
 @dataclass(init=False, unsafe_hash=True)
 class CodedLocation:
@@ -36,6 +39,64 @@ class CodedLocation:
 
         self._code = f"{self.lat:.{self.display_places}f}~{self.lon:.{self.display_places}f}"
 
+    def __lt__(self, other: "CodedLocation"):
+        """
+        Less-than comparator to enable sorting for CodedLocations.
+
+        For arithmetic comparisons we expect:
+
+        * north comes before south
+        * when on the same latitude, west comes before east
+
+        Note:
+            Coded locations at different resolutions are not considered
+            equal. Use `.as_tuple` to check for numerical equivalency.
+
+        Examples:
+            >>> origin = CodedLocation(-11.11, 111.11, resolution=0.001)
+            >>> origin == CodedLocation(-11.11, 111.11, resolution=0.001)  # Equal
+            True
+            >>> origin == CodedLocation(-11.11, 111.11, resolution=0.01)  # Wrong resolution
+            False
+
+            >>> origin < CodedLocation(-11.12, 111.11, resolution=0.001)  # South of origin
+            False
+            >>> origin < CodedLocation(-11.10, 111.11, resolution=0.001)  # North of origin
+            True
+            >>> origin < CodedLocation(-11.11, 111.10, resolution=0.001)  # West of origin
+            False
+            >>> origin < CodedLocation(-11.11, 111.12, resolution=0.001)  # East of origin
+            True
+
+        """
+        lat_delta = self.lat - other.lat
+        lon_delta = self.lon - other.lon
+        if lat_delta > 0:
+            return True
+        elif lat_delta < 0:
+            return False
+        elif lon_delta < 0:
+            return True
+        else:
+            return False
+
+    @property
+    def as_tuple(self) -> LatLon:
+        """
+        Convert to a `LatLon(latitude, longitude)` named tuple.
+
+        Example:
+            ```py
+            >>> from nzshm_common import location
+            >>> location.get_locations(["CHC"])[0]
+            CodedLocation(lat=-43.53, lon=172.63, resolution=0.001)
+            >>> latitude, longitude = location.get_locations(["CHC"])[0].as_tuple
+            >>> latitude
+            -43.53
+            ```
+        """
+        return LatLon(self.lat, self.lon)
+
     @property
     def code(self) -> str:
         """
@@ -43,10 +104,52 @@ class CodedLocation:
         """
         return self._code
 
+    @classmethod
+    def from_tuple(cls, location: LatLon, resolution: float = DEFAULT_RESOLUTION) -> "CodedLocation":
+        """
+        Create a `CodedLocation` from a tuple.
+
+        Parameters:
+            location: a structure containing a latitude and longitude, in that order
+            resolution: coordinate resolution in degrees
+
+        Examples:
+            Convert a single location:
+            >>> from nzshm_common import CodedLocation
+            >>> CodedLocation.from_tuple((-36.87, 174.77))
+            CodedLocation(lat=-36.87, lon=174.77, resolution=0.001)
+
+            >>> from nzshm_common import LatLon
+            >>> CodedLocation.from_tuple(LatLon(-36.87, 174.77))
+            CodedLocation(lat=-36.87, lon=174.77, resolution=0.001)
+
+            Convert a list of locations:
+            >>> location_list = [(-36.111, 174.111), (-36.222, 174.222)]
+            >>> list(map(CodedLocation.from_tuple, location_list))
+            [
+                CodedLocation(lat=-36.111, lon=174.111, resolution=0.001),
+                CodedLocation(lat=-36.222, lon=174.222, resolution=0.001)
+            ]
+
+            Convert a list of locations with a custom resolution:
+            >>> from functools import partial
+            >>> lo_res = partial(CodedLocation.from_tuple, resolution=0.1)
+            >>> list(map(lo_res, location_list))
+            [
+                CodedLocation(lat=-36.1, lon=174.1, resolution=0.1),
+                CodedLocation(lat=-36.2, lon=174.2, resolution=0.1)
+            ]
+        """
+        return CodedLocation(lat=location[0], lon=location[1], resolution=resolution)
+
     def resample(self, resolution: float) -> "CodedLocation":
-        """Downsample/Resample."""
+        """
+        Create a resampled CodedLocation with a finer resolution.
+        """
         return self.downsample(resolution)
 
     def downsample(self, resolution: float) -> "CodedLocation":
-        """Downsample/Resample."""
+        """
+        Create a downsampled CodedLocation with a coarser resolution.
+        """
         return CodedLocation(lat=self.lat, lon=self.lon, resolution=resolution)